improve numbers doc snippets
