+++ apps/docs-snippets2/src/types/numbers/numbers-2.ts
@@ -1,0 +1,10 @@
+// #region snippet-1
+import { bn } from 'fuels';
+
+const strNumber = '9007199254740992';
+
+const bigNumber = bn(strNumber);
+
+console.log('equals', bigNumber.toString() === strNumber);
+// true
+// #endregion snippet-1
