+++ apps/docs-snippets2/src/types/numbers/numbers-3.ts
@@ -1,0 +1,18 @@
+import { bn, Provider, Wallet } from 'fuels';
+
+import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../env';
+import { EchoValuesFactory } from '../typegend';
+
+const provider = await Provider.create(LOCAL_NETWORK_URL);
+const wallet = Wallet.fromPrivateKey(WALLET_PVT_KEY, provider);
+const deploy = await EchoValuesFactory.deploy(wallet);
+const { contract } = await deploy.waitForResult();
+
+// #region snippet-1
+const bigNumber = bn('1000000000000000000000000');
+
+const { value } = await contract.functions.echo_u64(bigNumber).get();
+
+console.log('value', value.toString());
+// '1000000000000000000000000'
+// #endregion snippet-1
